refactor(dev-tool): rename helperCompiler.ts to helpers.ts
Simpler, clearer module name.
